requirements.txt
amqp==5.1.1
asgiref==3.6.0
async-timeout==4.0.2
bcrypt==4.0.1
billiard==3.6.4.0
bleach==6.0.0
celery==5.2.7
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==3.1.0
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
cron-descriptor==1.2.35
cryptography==39.0.2
Django==4.1.7
django-celery-beat==2.5.0
django-crispy-forms==2.0
django-extensions==3.2.1
django-label-tag-attr==1.0.2
django-multiselectfield==0.1.12
django-resized==1.0.2
django-simple-history==3.3.0
django-summernote==0.8.20.0
django-timezone-field==5.0
djangorestframework==3.14.0
ecdsa==0.18.0
et-xmlfile==1.1.0
gunicorn==20.1.0
idna==3.4
jsonfield==3.1.0
kombu==5.2.4
mysqlclient==2.1.1
numpy==1.24.2
openpyxl==3.1.2
pandas==1.5.3
passlib==1.7.4
Pillow==9.4.0
plotly==5.13.1
prompt-toolkit==3.0.38
pyasn1==0.4.8
pycparser==2.21
PyJWT==2.6.0
python-crontab==2.7.1
python-dateutil==2.8.2
python-dotenv==1.0.0
python-jose==3.3.0
pytz==2022.7.1
redis==4.5.1
requests==2.28.2
rsa==4.9
six==1.16.0
sqlparse==0.4.3
tenacity==8.2.2
twilio==7.16.5
tzdata==2022.7
urllib3==1.26.15
vine==5.0.0
wcwidth==0.2.6
webencodings==0.5.1
WooCommerce==3.0.0
XlsxWriter==3.0.9
